重构和新增功能实现
在 `L34.h` 中将部分变量类型改为 `size_t`，并进行类型转换。重命名 `LCR170.h` 中的类名为 `BITA`，并添加更新和查询方法。更新项目文件以包含新添加的 `L327` 和 `L406` 相关文件。在 `common.h` 中添加必要的头文件。在 `main.cpp` 中实例化并调用新类的方法。新增 `L327` 和 `L406` 类，分别实现范围和计算和队列重建功能。

+++ Lee/L34.h
@@ -6,12 +6,12 @@ class L34
 {
 public:
     int lower_bound(vector<int>& nums, int target) {
-        int l = 0,r=nums.size()-1;
+        size_t l = 0,r=nums.size()-1;
 
-        int ans = -1;
+        size_t ans = -1;
 
         while (l <= r) {
-            int mid = l + (r - l) / 2;
+            size_t mid = l + (r - l) / 2;
             if (nums.at(mid) == target) {
                 ans = mid;
             }
@@ -23,17 +23,17 @@ public:
             }
         }
 
-        return ans;
+        return static_cast<int>(ans);
     }
 
 
     int upper_bound(vector<int>& nums, int target) {
-        int l = 0, r = nums.size() - 1;
+        size_t l = 0, r = nums.size() - 1;
 
-        int ans = -1;
+        size_t ans = -1;
 
         while (l <= r) {
-            int mid = l + (r - l) / 2;
+            size_t mid = l + (r - l) / 2;
             if (nums.at(mid) == target) {
                 ans = mid;
             }
@@ -45,7 +45,7 @@ public:
             }
         }
 
-        return ans;
+        return static_cast<int>(ans);
     }
 
     vector<int> searchRange(vector<int>& nums, int target) {
